correção do form do feed, que atualizava o render ao digitar
